Linked Server Tester
SQL Code to Test all Linked Servers that exist on a SQL Instance or group of CMS Instances.

Note this code reduces the connection timeout of each linked server before testing them then reverts them all back to 0 (Future improvements revert it back to the original value). This is to reduce the time spent waiting for bad connections to time-out.